feat(parse): preprocess links and paths
Relative paths are replaced by relative paths from the `src` path
and links to internal headers from the same file or a different
file are replaced by the correct and unique assigned id. Also,
links to files of the book are replaced by a link to the title
header of the file.

A warning is displayed if a link to a header does not exist
(wrong header or header with level greater than 6).